requirements.txt
plyer==2.1.0
selenium==3.141.0
urllib3==1.26.20
requestscertifi==2021.10.8
chardet==4.0.0
idna==2.10
python-dotenv==1.1.1
requests==2.27.1
typing==3.10.0.0
